feat: add multi select mode for SubsetSelect
This component does not use v-btn-toggle or another kind of
v-item-group as might be expected, because of the complexity
involved in switching between single and multi select. When
switching, the type of v-model changes from integer to array and
gets out of sync with what is displayed in the GUI. Also, there's
different logic needed to display what is selected based on the
mode.