fix: prevent color picker from entering an update loop
This commit uses a traitlet for updating the layer color and
immediately updating the frontend. A debounce is used for updating
the glue-state so it does not have to process all intermediate
color changes, which can start to back up or enter an update loop
when many intermediate changes come in, or the layer processing takes long.

Fixes part 2 of #312