ci: pin playwright similar to solara for compatibility
As mentioned in https://github.com/widgetti/solara/issues/913
pytest-ipywidgets is not compatible with newer versions of
pytest-playwright.
We did not notice this in solara, because we pinned playwright, which
seems to limit the versions of pytest-playwright to a compatible one.
If we apply the same pinning as solara, we should be able to avoid
this issue.